requirements.txt
flask==3.0.3
SQLAlchemy==2.0.35
PyMySQL==1.1.1
python-dotenv==1.0.1
cryptography>=41